Adding code for a new predictor : agree_bimode_combined
This predictor is based on the combination of bimode and agree
predictors. It essentially use bimode type 2 PHTs along with a CPT and
it updates the counters in a similar manner. It however also adds a bias
to each branch, thus each PHT counter now either agrees with the
bias(TAKEN) or it disagrees(NOT TAKEN).